pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pyBELLA"
version = "0.5.0"
dependencies = [
    "numba",
    "numpy==1.22.1",
    "h5py",
    "scipy",
    "PyYAML",
    "termcolor",
    "dask",
    "dask[distributed]",
    "matplotlib",
    "dill"
]

[tool.setuptools]
package-dir = {"pybella" = "src"}